Don't update git repos from scratch
9c913b93e275038393e4ba5f67b7421919d93cf9 changed the behavior of
git_repository to unconditionally delete the existing directory and
refetch every time (previously the directory was only deleted if the
.git/ directory was not present). This results in having to perform a
fetch on every build, which is either unnecessary (in the case that the
desired ref is a commit, and that commit is already checked out), or
unnecessarily expensive (because any previously fetched data that could
reduce the amount of new data that needs to be downloaded over the
network is discarded).

This commit removes the deletion of the existing repo directory, but
maintains the desired clean state by making add_origin idempotent and
proactively removing any submodules that might exist.